Performed Networking from JSON

+++ Kursvalut/Model/Currency.swift
@@ -2,9 +2,8 @@
 import Foundation
 
 struct Currency {
-    
     var fullName: String
     var shortName: String
-    var currentValue: Float
-    var previousValue: Float
+    var currentValue: Double
+    var previousValue: Double
 }
